pyproject.toml
[project]
name = "prism-ai"
version = "1.0.0"
description = "PRism-AI: Limitless, self-evolving, free & open-source AI-powered PR review agent"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
authors = [{ name = "WhoisMonesh" }]
keywords = ["pr-review", "ai", "devops", "ollama", "vertex-ai", "bedrock", "llm", "github", "gitlab", "gitea", "open-source"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "starlette>=0.40.0",
    # HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Config
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
    "dynaconf>=3.2.5",
    "toml>=0.10.2",
    # Auth & Security
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    # LLM unified gateway
    "litellm>=1.40.0",
    "tiktoken>=0.7.0",
    "tenacity>=8.3.0",
    # Logging
    "loguru>=0.7.2",
    # Git providers
    "PyGithub>=2.3.0",
    "python-gitlab>=4.5.0",
    "gitpython>=3.1.43",
    # Cloud - GCP
    "google-cloud-aiplatform>=1.57.0",
    "google-auth>=2.29.0",
    "google-cloud-secret-manager>=2.20.0",
    # Cloud - AWS
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    # Templates
    "Jinja2>=3.1.4",
    # Scheduling / Evolution
    "apscheduler>=3.10.4",
    "schedule>=1.2.2",
    # Metrics
    "prometheus-client>=0.20.0",
    # Utilities
    "rich>=13.7.0",
    "click>=8.1.7",
    "ujson>=5.9.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx",
    "faker>=25.0.0",
]
vault = ["hvac>=2.1.0"]
bitbucket = ["atlassian-python-api>=3.41.0"]
azure = ["azure-devops>=7.1.0", "azure-identity>=1.16.0"]
all = ["prism-ai[dev,vault,bitbucket,azure]"]

[project.scripts]
prism-ai = "src.cli:main"

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__init__.py"]
